backend/requirements.txt
fastapi
uvicorn[standard]
pillow
torch
timm
torchvision
python-multipart
